+++ Assets/Mirror/Runtime/Transport/Telepathy/Server.cs
@@ -145,8 +145,7 @@ namespace Telepathy
                             // otherwise the send thread would only end if it's
                             // actually sending data while the connection is
                             // closed.
-                            // => AbortAndJoin is the safest way and avoids race conditions!
-                            sendThread.AbortAndJoin();
+                            sendThread.Interrupt();
                         }
                         catch (Exception exception)
                         {
@@ -215,18 +214,8 @@ namespace Telepathy
 
             // kill listener thread at all costs. only way to guarantee that
             // .Active is immediately false after Stop.
-            // => AbortAndJoin is the safest way and avoids race conditions!
-            listenerThread?.AbortAndJoin();
-
-            // wait until thread is TRULY finished. this is the only way
-            // to guarantee that everything was properly cleaned up before
-            // returning.
-            // => this means that calling Stop() may sometimes block
-            //    for a while, but there is no other way to guarantee that
-            //    everything is cleaned up properly by the time Stop() returns.
-            //    we have to live with the wait time.
-            listenerThread?.Join();
-
+            // -> calling .Join would sometimes wait forever
+            listenerThread?.Interrupt();
             listenerThread = null;
 
             // close all client connections